# Conflicts:
#	app/models/admin_activity_change_logger.rb
#	app/models/client_kyc_config_detail.rb
#	app/services/authentication/api_request/base.rb
#	app/services/client_management/reset_secret_key.rb
#	config/routes/saas_api/v2.rb
#	lib/global_constant/admin_activity_change_logger.rb
#	lib/tasks/onetimer/add_client.rake

+++ app/controllers/web/admin/base_controller.rb
@@ -1,5 +1,7 @@
 class Web::Admin::BaseController < Web::WebController
 
+  include Util::ResultHelper
+
   before_action :authenticate_request
 
   private
@@ -10,14 +12,13 @@ class Web::Admin::BaseController < Web::WebController
   # * Date: 10/10/2017
   # * Reviewed By: Sunil Khedar
   #
-  def authenticate_request(is_super_admin_role=false)
+  def authenticate_request(options = {is_super_admin_role: false, validate_terms_of_use: true})
 
     service_response = AdminManagement::VerifyCookie::DoubleAuth.new(
         cookie_value: cookies[GlobalConstant::Cookie.admin_cookie_name.to_sym],
         browser_user_agent: http_user_agent,
-        is_super_admin_role: is_super_admin_role
+        options: options
     ).perform
-
     if service_response.success?
       # Update Cookie, if required
       extended_cookie_value = service_response.data[:extended_cookie_value]
@@ -33,10 +34,79 @@ class Web::Admin::BaseController < Web::WebController
       # Remove sensitive data
       service_response.data = {}
     else
-      delete_cookie(GlobalConstant::Cookie.admin_cookie_name)
+      if service_response.http_code == GlobalConstant::ErrorCode.unauthorized_access
+        delete_cookie(GlobalConstant::Cookie.admin_cookie_name)
+      end
       render_api_response(service_response)
     end
 
   end
+
+  # Check if single auth logged in admin and redirect to required page
+  #
+  # * Author: Aman
+  # * Date: 05/02/2019
+  # * Reviewed By:
+  #
+  def is_single_auth_logged_in
+    auth_cookie = cookies[GlobalConstant::Cookie.admin_cookie_name.to_sym]
+    return if auth_cookie.blank?
+
+    service_response = AdminManagement::VerifyCookie::SingleAuth.new(
+        cookie_value: auth_cookie,
+        browser_user_agent: http_user_agent
+    ).perform
+
+
+    if service_response.success?
+      err = error_with_internal_code('c_w_a_bc_vsali_1',
+                                     'Redirecting',
+                                     GlobalConstant::ErrorCode.temporary_redirect,
+                                     {},
+                                     {},
+                                     {}
+      )
+
+      err.set_error_extra_info({redirect_url: GlobalConstant::WebUrls.multifactor_auth})
+      render_api_response(err) and return
+    else
+      # delete_cookie(GlobalConstant::Cookie.admin_cookie_name)
+    end
+
+  end
+
+  # Check if double_auth logged in admin and redirect to required page
+  #
+  # * Author: Aman
+  # * Date: 05/02/2019
+  # * Reviewed By:
+  #
+  def is_double_auth_logged_in
+    auth_cookie = cookies[GlobalConstant::Cookie.admin_cookie_name.to_sym]
+    return if auth_cookie.blank?
+
+    service_response = AdminManagement::VerifyCookie::DoubleAuth.new(
+        cookie_value: auth_cookie,
+        browser_user_agent: http_user_agent,
+        options: {is_super_admin_role: false, validate_terms_of_use: true}
+    ).perform
+
+    if service_response.http_code == GlobalConstant::ErrorCode.temporary_redirect
+      render_api_response(service_response)
+    elsif service_response.success?
+      err = error_with_internal_code('c_w_a_bc_vdali_1',
+                                     'Redirecting',
+                                     GlobalConstant::ErrorCode.temporary_redirect,
+                                     {},
+                                     {},
+                                     {}
+      )
+      err.set_error_extra_info({redirect_url: GlobalConstant::WebUrls.admin_dashboard})
+      render_api_response(err)
+    else
+      # delete_cookie(GlobalConstant::Cookie.admin_cookie_name)
+    end
+  end
+
 
 end
